requirements.txt
boto3
xmltodict
python-dotenv
gc-python-utils